Move migrated contents in the proper controller
